Move components to individual files
