+++ lib/drag-and-drop/Area.jsx
@@ -1,7 +1,8 @@
 import { createSignal } from 'solid-js';
 import { createStore } from 'solid-js/store';
+import Column from './Column.jsx';
 
-export const Area = (props) => {
+export default (props) => {
   const [itemIndex, setItemIndex] = createSignal(0);
   const [colIndex, setColIndex] = createSignal(0);
   const [items, setItems] = createStore([]);
@@ -82,59 +83,5 @@ export const Area = (props) => {
         </For>
       </div>
     </>
-  );
-};
-
-export const Column = (props) => {
-  const items = () => {
-    return props.items.filter((item) => item.colID == props.colID);
-  };
-
-  const handleDrop = (e) => {
-    console.log('dropping');
-    e.preventDefault();
-    const itemID = e.dataTransfer.getData('id');
-    props.updateItems('update', { itemID, colID: props.colID });
-  };
-
-  return (
-    <div
-      class='border border-black m-3 p-3'
-      onDrop={(e) => handleDrop(e)}
-      onDragOver={(e) => e.preventDefault()}
-    >
-      column
-      <button
-        class='border border-black m-3 p-3'
-        onClick={() =>
-          props.updateItems('create', {
-            title: 'new item',
-            content: 'some content',
-            colID: props.colID,
-          })
-        }
-      >
-        New Item
-      </button>
-      <For each={items()}>
-        {(item) => <Item itemID={item.itemID} {...props} />}
-      </For>
-    </div>
-  );
-};
-
-export const Item = (props) => {
-  const handleDragStart = (e) => {
-    e.dataTransfer.setData('id', props.itemID);
-  };
-
-  return (
-    <div
-      class='border border-black m-3 p-3 cursor-move'
-      draggable={true}
-      onDragStart={(e) => handleDragStart(e)}
-    >
-      item {props.itemID}
-    </div>
   );
 };
